+++ inst/python/kerastools/callback.py
@@ -1,5 +1,4 @@
 
-
 import os
 
 if (os.getenv('KERAS_IMPLEMENTATION', 'tensorflow') == 'keras'):
@@ -50,50 +49,82 @@ class RCallback(Callback):
     self._chief_worker_only = False
 
   def on_epoch_begin(self, epoch, logs=None):
-    self.r_on_epoch_begin(epoch, logs)
+    out = self.r_on_epoch_begin(epoch, logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)
 
   def on_epoch_end(self, epoch, logs=None):
-    self.r_on_epoch_end(epoch, logs)
+    out = self.r_on_epoch_end(epoch, logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)
 
   def on_train_begin(self, logs=None):
     self.r_set_context(self.params, self.model)
-    self.r_on_train_begin(logs)
+    out = self.r_on_train_begin(logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)
 
   def on_train_end(self, logs=None):
-    self.r_on_train_end(logs)
+    out = self.r_on_train_end(logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)
 
   def on_batch_begin(self, batch, logs=None):
-    self.r_on_batch_begin(batch, logs)
+    out = self.r_on_batch_begin(batch, logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)
 
   def on_batch_end(self, batch, logs=None):
-    self.r_on_batch_end(batch, logs)
+    out = self.r_on_batch_end(batch, logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)
 
   def on_predict_batch_begin(self, batch, logs=None):
-    self.r_on_predict_batch_begin(batch, logs)
+    out = self.r_on_predict_batch_begin(batch, logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)
 
   def on_predict_batch_end(self, batch, logs=None):
-    self.r_on_predict_batch_end(batch, logs)
+    out = self.r_on_predict_batch_end(batch, logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)
 
   def on_predict_begin(self, logs=None):
-    self.r_on_predict_begin(logs)
+    out = self.r_on_predict_begin(logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)
 
   def on_predict_end(self, logs=None):
-    self.r_on_predict_end(logs)
+    out = self.r_on_predict_end(logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)
 
   def on_test_batch_begin(self, batch, logs=None):
-    self.r_on_test_batch_begin(batch, logs)
+    out = self.r_on_test_batch_begin(batch, logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)
 
   def on_test_batch_end(self, batch, logs=None):
-    self.r_on_test_batch_end(batch, logs)
+    out = self.r_on_test_batch_end(batch, logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)
 
   def on_test_begin(self, logs=None):
-    self.r_on_test_begin(logs)
+    out = self.r_on_test_begin(logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)
 
   def on_test_end(self, logs=None):
-    self.r_on_test_end(logs)
+    out = self.r_on_test_end(logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)
 
   def on_train_batch_begin(self, batch, logs=None):
-    self.r_on_train_batch_begin(batch, logs)
+    out = self.r_on_train_batch_begin(batch, logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)
 
   def on_train_batch_end(self, batch, logs=None):
-    self.r_on_train_batch_end(batch, logs)
+    out = self.r_on_train_batch_end(batch, logs)
+    if isinstance(out, dict) and isinstance(logs, dict):
+      logs.update(out)